requirements.txt
abnf==2.2.0
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
airportsdata==20250909
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
astor==0.8.1
async-timeout==5.0.1
attrs==25.4.0
backoff==2.2.1
bitsandbytes==0.49.1
blake3==1.0.8
blobfile==3.1.0
cachetools==6.2.4
cbor2==5.8.0
certifi==2026.1.4
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
cint==1.0.0
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.10.2
cryptography==46.0.3
cupy-cuda12x==13.6.0
cut-cross-entropy==25.1.1
datasets==4.3.0
depyf==0.19.0
diffusers==0.36.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.1
email-validator==2.3.0
exceptiongroup==1.3.1
Farama-Notifications==0.0.4
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.10.1
fastar==0.8.0
fastrlock==0.8.3
fickling==0.1.7
filelock==3.20.3
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.9.0
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.46
gql==4.0.0
graphql-core==3.2.7
graphviz==0.21
gymnasium==0.29.0
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
interegular==0.3.3
intervaltree==3.2.1
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kagglehub==0.4.0
kagglesdk==0.1.14
kaitaistruct==0.11
lark==1.2.2
litellm==1.74.1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.12
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
miniwob==1.0
mistral_common==1.8.8
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
openai==1.90.0
openai-harmony==0.0.8
opencv-python-headless==4.11.0.86
openpipe-art==0.5.4
outcome==1.3.0.post0
outlines==0.0.46
outlines_core==0.2.10
packaging==25.0
pandas==2.3.3
partial-json-parser==0.2.1.1.post7
pdfminer.six==20260107
peft==0.18.1
pillow==12.1.0
platformdirs==4.5.1
polars==1.37.1
polars-runtime-32==1.37.1
polyfile-weave==0.5.8
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.0
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
py-cpuinfo==9.0.0
pyairports==0.0.1
pyarrow==22.0.0
pybase64==1.4.3
pycountry==24.6.1
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ray==2.53.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rpds-py==0.30.0
safetensors==0.7.0
scipy==1.15.3
selenium==4.39.0
sentencepiece==0.2.1
sentry-sdk==2.49.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==1.0.0
starlette==0.50.0
stdlib-list==0.11.1
sympy==1.14.0
tblib==3.2.2
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.7.1
torchao==0.15.0
torchaudio==2.7.1
torchdata==0.11.0
torchtune==0.6.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.57.3
trio==0.32.0
trio-websocket==0.12.2
triton==3.3.1
trl==0.24.0
typeguard==4.4.4
typer==0.21.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==1.0.5
tzdata==2025.3
unsloth==2025.12.9
unsloth_zoo==2025.12.7
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
vllm==0.10.0
wandb==0.24.0
watchfiles==1.1.1
weave==0.52.23
websocket-client==1.9.0
websockets==16.0
wsproto==1.3.2
xformers==0.0.31
xgrammar==0.1.21
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
